工作流脚本/同人提取/requirements.txt
fastapi
uvicorn
jieba
numpy
faiss-cpu
sentence-transformers
python-dotenv
requests
pydantic
openai
python-multipart
pytest
